added the songlist and removed the now unnecessary stuff
