+++ app/src/main/java/com/hich2000/tagcapella/MainActivity.kt
@@ -12,24 +12,14 @@ import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowRightAlt
-import androidx.compose.material.icons.filled.Pause
-import androidx.compose.material.icons.filled.PlayArrow
-import androidx.compose.material.icons.filled.Repeat
-import androidx.compose.material.icons.filled.RepeatOne
-import androidx.compose.material.icons.filled.Shuffle
-import androidx.compose.material.icons.filled.ShuffleOn
-import androidx.compose.material.icons.filled.SkipNext
-import androidx.compose.material.icons.filled.SkipPrevious
-import androidx.compose.material3.BottomAppBar
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -39,11 +29,20 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
-import androidx.media3.common.Player
+import androidx.media3.common.MediaItem
+import androidx.media3.common.MediaMetadata
+import com.hich2000.tagcapella.music_player.MusicControls
 import com.hich2000.tagcapella.music_player.MusicPlayerViewModel
+import com.hich2000.tagcapella.music_player.SongCard
 import com.hich2000.tagcapella.theme.TagcapellaTheme
+import kotlin.io.path.Path
+import kotlin.io.path.isDirectory
+import kotlin.io.path.isRegularFile
+import kotlin.io.path.listDirectoryEntries
+import kotlin.io.path.nameWithoutExtension
 
 class MainActivity : ComponentActivity() {
 
@@ -107,19 +106,56 @@ class MainActivity : ComponentActivity() {
         // Use the state variable to determine if the MediaController is initialized
         val isMediaControllerInitialized by mediaViewModel.isMediaControllerInitialized
 
+
+        val playlist = mutableListOf<MediaItem>()
+        val path = Path("/storage/emulated/0/Music").listDirectoryEntries()
+        path.listIterator().forEach {
+            if (!it.isDirectory() && it.isRegularFile()) {
+
+                val mediaItem = MediaItem.Builder()
+                    .setMediaId(it.toString())
+                    .setUri(it.toString())
+                    .setMediaMetadata(
+                        MediaMetadata.Builder()
+                            .setTitle(it.nameWithoutExtension)
+                            .setDisplayTitle(it.nameWithoutExtension)
+                            .build()
+                    )
+                    .build()
+
+                playlist.add(mediaItem)
+            }
+        }
+
+
         CompositionLocalProvider(LocalMusicPlayerViewModel provides mediaViewModel) {
             if (isMediaControllerInitialized) {
                 Scaffold(
                     modifier = Modifier.fillMaxSize(),
                     bottomBar = {
                         MusicControls()
+                    },
+                    topBar = {
+                        Text(
+                            "ayylmao",
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .border(2.dp, Color.Gray),
+                            textAlign = TextAlign.Center
+                        )
                     }
                 ) { innerPadding ->
-                    Text(
-                        text = "Media Controller Initialized!",
+                    Column(
                         modifier = Modifier
+                            .fillMaxWidth()
+                            .fillMaxHeight()
                             .padding(innerPadding)
-                    )
+                            .verticalScroll(rememberScrollState(0))
+                    ) {
+                        for ((index, song) in playlist.withIndex()) {
+                            SongCard(song, index)
+                        }
+                    }
                 }
             } else {
                 CircularProgressIndicator(
@@ -127,111 +163,5 @@ class MainActivity : ComponentActivity() {
                 )
             }
         }
-    }
-}
-
-@Composable
-fun MusicControls() {
-    //get the mediaController for controls
-    val mediaController = LocalMusicPlayerViewModel.current.mediaController
-    //observe the isPlaying state for ui changes
-    val isPlaying by LocalMusicPlayerViewModel.current.isPlaying
-    //observe the shuffleModeEnabled state for ui changes
-    val shuffleModeEnabled by LocalMusicPlayerViewModel.current.shuffleModeEnabled
-    //observe the loopMode state for ui changes
-    val repeatMode by LocalMusicPlayerViewModel.current.repeatMode
-
-
-    BottomAppBar(
-        modifier = Modifier
-            .border(2.dp, Color.Gray)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .border(2.dp, Color.Gray),
-            horizontalArrangement = Arrangement.Center
-        ) {
-            //shuffle mode
-            IconButton(
-                onClick = {
-                    if (shuffleModeEnabled) {
-                        mediaController.shuffleModeEnabled = false
-                    } else {
-                        mediaController.shuffleModeEnabled = true
-                    }
-                }
-            ) {
-                val icon = if (shuffleModeEnabled) Icons.Default.ShuffleOn else Icons.Default.Shuffle
-                Icon(
-                    icon,
-                    contentDescription = "Shuffle button"
-                )
-            }
-            //skip previous
-            IconButton(
-                onClick = {
-                    mediaController.seekToPrevious()
-                }
-            ) {
-                Icon(
-                    Icons.Default.SkipPrevious,
-                    contentDescription = "Skip to previous button"
-                )
-            }
-            //play/pause
-            IconButton(
-                onClick = {
-                    if (isPlaying) mediaController.pause() else mediaController.play()
-                }
-            ) {
-                val icon = if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow
-                val contentDescription = if (isPlaying) "Pause" else "Play"
-                Icon(
-                    icon,
-                    contentDescription = contentDescription
-                )
-            }
-            //skip next
-            IconButton(
-                onClick = {
-                    mediaController.seekToNext()
-                }
-            ) {
-                Icon(
-                    Icons.Default.SkipNext,
-                    contentDescription = "Skip to next button"
-                )
-            }
-            //loop mode
-            IconButton(
-                onClick = {
-                    if (repeatMode == Player.REPEAT_MODE_OFF) {
-                        mediaController.repeatMode = Player.REPEAT_MODE_ALL
-                    } else if (repeatMode == Player.REPEAT_MODE_ALL) {
-                        mediaController.repeatMode = Player.REPEAT_MODE_ONE
-                    } else if (repeatMode == Player.REPEAT_MODE_ONE) {
-                        mediaController.repeatMode = Player.REPEAT_MODE_OFF
-                    }
-                }
-            ) {
-                var icon = Icons.Default.Repeat
-
-                if (repeatMode == Player.REPEAT_MODE_OFF) {
-                    icon = Icons.AutoMirrored.Filled.ArrowRightAlt
-                } else if (repeatMode == Player.REPEAT_MODE_ALL) {
-                    icon = Icons.Default.Repeat
-                } else if (repeatMode == Player.REPEAT_MODE_ONE) {
-                    icon = Icons.Default.RepeatOne
-                }
-
-                Icon(
-                    icon,
-                    contentDescription = "Shuffle button"
-                )
-            }
-
-        }
-
     }
 }
